extract is just weighted event:tag tuples
This is to dodge an error in MariaDB's merging of CTEs. Extracting the
three parts (events, tags, and uses) with three SELECT statements gave
the right result, but when they're UNIONed, the merge does something
stupid and ends up providing extra rows in the tags part (from events
which are excluded in the CTE).